requirements.txt
nltk==3.8.1
streamlit==1.28.2
word2number==1.1
regex==2023.10.3
requests==2.31.0
pandas==2.1.1
urllib3==1.26.5
beautifulsoup4==4.9.3
fuzzywuzzy==0.18.0
tabulate==0.9.0
inflect==7.0.0
scikit-surprise==1.1.3
scikit-learn==1.3.1
transformers==4.34.1
torch==2.1.0

